test: consolidate secrets integration coverage
