--- plugins/rust/python-package/secrets_detection/tests/mcpgateway_mock/__init__.py
@@ -1,1 +1,0 @@
-"""Mock mcpgateway package for plugin tests."""
