--- plugins/rust/python-package/secrets_detection/tests/mcpgateway_mock/plugins/__init__.py
@@ -1,1 +1,0 @@
-"""Mock mcpgateway.plugins package for plugin tests."""
