--- plugins/rust/python-package/secrets_detection/tests/test_plugin.py
@@ -1,910 +1,0 @@
-import subprocess
-import sys
-
-import pytest
-from pydantic import BaseModel, RootModel, model_serializer
-
-from mcpgateway_mock.plugins.framework import (
-    GlobalContext,
-    PluginConfig,
-    PluginContext,
-    PromptPrehookPayload,
-    ResourceContent,
-    ResourcePostFetchPayload,
-    ToolPostInvokePayload,
-)
-
-from cpex_secrets_detection.secrets_detection import SecretsDetectionPlugin
-from cpex_secrets_detection.secrets_detection_rust import py_scan_container
-
-
-def _make_config(**overrides) -> PluginConfig:
-    config = {
-        "block_on_detection": False,
-        "redact": True,
-        "redaction_text": "[REDACTED]",
-    }
-    config.update(overrides)
-    return PluginConfig(name="secrets_detection", config=config)
-
-
-def _make_context() -> PluginContext:
-    return PluginContext(global_context=GlobalContext(user="user-1"))
-
-
-class TestPluginHooks:
-    @pytest.fixture
-    def plugin(self):
-        return SecretsDetectionPlugin(_make_config())
-
-    async def test_prompt_pre_fetch_redacts_without_blocking(self, plugin):
-        payload = PromptPrehookPayload(
-            prompt_id="prompt-1",
-            args={"input": "AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE"},
-        )
-
-        result = await plugin.prompt_pre_fetch(payload, _make_context())
-
-        assert result.continue_processing is True
-        assert result.violation is None
-        assert result.modified_payload is not None
-        assert result.modified_payload.args["input"] == "AWS_ACCESS_KEY_ID=[REDACTED]"
-        assert result.metadata == {"secrets_redacted": True, "count": 1}
-
-    async def test_prompt_pre_fetch_leaves_clean_payload_unmodified(self, plugin):
-        payload = PromptPrehookPayload(
-            prompt_id="prompt-1",
-            args={"input": "hello world"},
-        )
-
-        result = await plugin.prompt_pre_fetch(payload, _make_context())
-
-        assert result.continue_processing is True
-        assert result.violation is None
-        assert result.modified_payload is None
-        assert result.metadata == {}
-
-    async def test_prompt_pre_fetch_blocks_without_redaction(self):
-        plugin = SecretsDetectionPlugin(_make_config(block_on_detection=True, redact=False))
-        payload = PromptPrehookPayload(
-            prompt_id="prompt-1",
-            args={"input": "AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE"},
-        )
-
-        result = await plugin.prompt_pre_fetch(payload, _make_context())
-
-        assert result.continue_processing is False
-        assert result.violation is not None
-        assert result.violation.code == "SECRETS_DETECTED"
-        assert result.modified_payload == payload
-
-    async def test_prompt_pre_fetch_blocks_with_redaction_without_leaking_secret(self):
-        plugin = SecretsDetectionPlugin(_make_config(block_on_detection=True, redact=True))
-        payload = PromptPrehookPayload(
-            prompt_id="prompt-1",
-            args={"input": "AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE"},
-        )
-
-        result = await plugin.prompt_pre_fetch(payload, _make_context())
-
-        assert result.continue_processing is False
-        assert result.violation is not None
-        assert result.violation.code == "SECRETS_DETECTED"
-        assert result.modified_payload is not None
-        assert result.modified_payload is not payload
-        assert result.modified_payload.args["input"] == "AWS_ACCESS_KEY_ID=[REDACTED]"
-        assert payload.args["input"] == "AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE"
-
-    async def test_prompt_pre_fetch_metadata_omits_match_previews(self):
-        plugin = SecretsDetectionPlugin(_make_config(redact=False))
-        payload = PromptPrehookPayload(
-            prompt_id="prompt-1",
-            args={"input": "AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE"},
-        )
-
-        result = await plugin.prompt_pre_fetch(payload, _make_context())
-
-        assert result.metadata is not None
-        assert result.metadata["count"] == 1
-        assert result.metadata["secrets_findings"] == [{"type": "aws_access_key_id"}]
-
-    async def test_prompt_pre_fetch_blocking_details_omit_match_previews(self):
-        plugin = SecretsDetectionPlugin(_make_config(block_on_detection=True, redact=False))
-        payload = PromptPrehookPayload(
-            prompt_id="prompt-1",
-            args={"input": "AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE"},
-        )
-
-        result = await plugin.prompt_pre_fetch(payload, _make_context())
-
-        assert result.violation is not None
-        assert result.violation.details == {
-            "count": 1,
-            "examples": [{"type": "aws_access_key_id"}],
-        }
-
-
-class TestPublicRustApi:
-    def test_scan_container_preserves_tuple_shape_when_clean(self):
-        payload = ("safe", 1, {"nested": "value"})
-
-        count, redacted, findings = py_scan_container(
-            payload, {"redact": True, "redaction_text": "[REDACTED]"}
-        )
-
-        assert count == 0
-        assert findings == []
-        assert redacted == payload
-        assert isinstance(redacted, tuple)
-
-    def test_scan_container_preserves_opaque_object_when_clean(self):
-        class SlotOnlyPayload:
-            __slots__ = ("value",)
-
-            def __init__(self, value):
-                self.value = value
-
-        payload = SlotOnlyPayload("safe")
-
-        count, redacted, findings = py_scan_container(
-            payload, {"redact": True, "redaction_text": "[REDACTED]"}
-        )
-
-        assert count == 0
-        assert findings == []
-        assert redacted is payload
-
-    def test_scan_container_redacts_custom_object_with_dict_state(self):
-        class SecretBox:
-            def __init__(self, value):
-                self.value = value
-
-        payload = SecretBox("AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE")
-
-        count, redacted, findings = py_scan_container(
-            payload, {"redact": True, "redaction_text": "[REDACTED]"}
-        )
-
-        assert count == 1
-        assert findings == [{"type": "aws_access_key_id"}]
-        assert redacted is not payload
-        assert isinstance(redacted, SecretBox)
-        assert redacted.value == "AWS_ACCESS_KEY_ID=[REDACTED]"
-        assert payload.value == "AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE"
-
-    def test_scan_container_omits_match_previews_from_public_findings(self):
-        count, _, findings = py_scan_container(
-            "AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE",
-            {"redact": True, "redaction_text": "[REDACTED]"},
-        )
-
-        assert count == 1
-        assert findings == [{"type": "aws_access_key_id"}]
-
-    def test_scan_container_prefers_redacted_serialized_view_when_both_paths_match(self):
-        class DualSurfacePayload:
-            def __init__(self):
-                self.state_secret = "AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE"
-
-            def model_dump(self):
-                return "AWS_SECRET_ACCESS_KEY=FAKESecretAccessKeyForTestingEXAMPLE0000"
-
-        payload = DualSurfacePayload()
-
-        count, redacted, findings = py_scan_container(
-            payload, {"redact": True, "redaction_text": "[REDACTED]"}
-        )
-
-        assert count == 2
-        assert findings == [
-            {"type": "aws_access_key_id"},
-            {"type": "aws_secret_access_key"},
-        ]
-        assert redacted == "[REDACTED]"
-
-    def test_scan_container_redacts_non_replayable_custom_object(self):
-        class NonReplayableBox:
-            def __init__(self, secret):
-                self.secret = secret
-                self.derived = "derived"
-
-        payload = NonReplayableBox("AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE")
-
-        count, redacted, findings = py_scan_container(
-            payload, {"redact": True, "redaction_text": "[REDACTED]"}
-        )
-
-        assert count == 1
-        assert len(findings) == 1
-        assert redacted is not payload
-        assert isinstance(redacted, NonReplayableBox)
-        assert redacted.secret == "AWS_ACCESS_KEY_ID=[REDACTED]"
-        assert redacted.derived == "derived"
-        assert payload.secret == "AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE"
-
-    def test_scan_container_redacts_slot_backed_custom_object(self):
-        class SlotSecretBox:
-            __slots__ = ("value",)
-
-            def __init__(self, value):
-                self.value = value
-
-        payload = SlotSecretBox("AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE")
-
-        count, redacted, findings = py_scan_container(
-            payload, {"redact": True, "redaction_text": "[REDACTED]"}
-        )
-
-        assert count == 1
-        assert len(findings) == 1
-        assert redacted is not payload
-        assert isinstance(redacted, SlotSecretBox)
-        assert redacted.value == "AWS_ACCESS_KEY_ID=[REDACTED]"
-        assert payload.value == "AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE"
-
-    def test_scan_container_redacts_hybrid_dict_and_slots_object(self):
-        class HybridSecretBox:
-            __slots__ = {"slot_secret": "slot", "__dict__": "dict"}
-
-            def __init__(self, slot_secret, label):
-                self.slot_secret = slot_secret
-                self.label = label
-
-        payload = HybridSecretBox(
-            "AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE",
-            "safe",
-        )
-
-        count, redacted, findings = py_scan_container(
-            payload, {"redact": True, "redaction_text": "[REDACTED]"}
-        )
-
-        assert count == 1
-        assert len(findings) == 1
-        assert redacted is not payload
-        assert isinstance(redacted, HybridSecretBox)
-        assert redacted.slot_secret == "AWS_ACCESS_KEY_ID=[REDACTED]"
-        assert redacted.label == "safe"
-        assert payload.slot_secret == "AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE"
-
-    def test_scan_container_redacts_guarded_object_without_running_setattr(self):
-        class GuardedSecretBox:
-            __slots__ = ("secret", "label", "_locked")
-
-            def __init__(self, secret, label):
-                object.__setattr__(self, "secret", secret)
-                object.__setattr__(self, "label", label)
-                object.__setattr__(self, "_locked", True)
-
-            def __setattr__(self, name, value):
-                raise AssertionError(f"unexpected setattr for {name}")
-                object.__setattr__(self, name, value)
-
-        payload = GuardedSecretBox(
-            "AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE",
-            "safe",
-        )
-
-        count, redacted, findings = py_scan_container(
-            payload, {"redact": True, "redaction_text": "[REDACTED]"}
-        )
-
-        assert count == 1
-        assert len(findings) == 1
-        assert redacted is not payload
-        assert isinstance(redacted, GuardedSecretBox)
-        assert redacted.secret == "AWS_ACCESS_KEY_ID=[REDACTED]"
-        assert redacted.label == "safe"
-        assert payload.secret == "AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE"
-
-    def test_scan_container_redacts_slots_declared_as_mapping(self):
-        class MappingSlotSecretBox:
-            __slots__ = {"secret": "slot doc"}
-
-            def __init__(self, secret):
-                self.secret = secret
-
-        payload = MappingSlotSecretBox("AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE")
-
-        count, redacted, findings = py_scan_container(
-            payload, {"redact": True, "redaction_text": "[REDACTED]"}
-        )
-
-        assert count == 1
-        assert len(findings) == 1
-        assert redacted is not payload
-        assert isinstance(redacted, MappingSlotSecretBox)
-        assert redacted.secret == "AWS_ACCESS_KEY_ID=[REDACTED]"
-        assert payload.secret == "AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE"
-
-    def test_scan_container_detects_secret_exposed_only_by_model_dump(self):
-        class SplitSecretModel(BaseModel):
-            prefix: str
-            suffix: str
-
-            @model_serializer(mode="plain")
-            def serialize_model(self):
-                return f"{self.prefix}{self.suffix}"
-
-        payload = SplitSecretModel(prefix="AKIA", suffix="FAKE12345EXAMPLE")
-
-        count, redacted, findings = py_scan_container(
-            payload, {"redact": True, "redaction_text": "[REDACTED]"}
-        )
-
-        assert count == 1
-        assert len(findings) == 1
-        assert redacted == "[REDACTED]"
-
-    def test_scan_container_redacts_secret_exposed_only_by_root_model_dump(self):
-        payload = RootModel[str]("AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE")
-
-        count, redacted, findings = py_scan_container(
-            payload, {"redact": True, "redaction_text": "[REDACTED]"}
-        )
-
-        assert count == 1
-        assert len(findings) == 1
-        assert isinstance(redacted, RootModel)
-        assert redacted.root == "AWS_ACCESS_KEY_ID=[REDACTED]"
-
-    def test_scan_container_handles_recursive_model_dump_without_crashing(self):
-        script = """
-from cpex_secrets_detection.secrets_detection_rust import py_scan_container
-
-class RecursivePayload:
-    def __init__(self):
-        self.secret = "AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE"
-
-    def model_dump(self):
-        return self
-
-count, redacted, findings = py_scan_container(
-    RecursivePayload(),
-    {"redact": True, "redaction_text": "[REDACTED]"},
-)
-assert count == 1
-assert len(findings) == 1
-assert redacted.secret == "AWS_ACCESS_KEY_ID=[REDACTED]"
-"""
-        result = subprocess.run(
-            [sys.executable, "-c", script],
-            capture_output=True,
-            text=True,
-            check=False,
-            timeout=5,
-        )
-
-        assert result.returncode == 0, result.stderr or result.stdout
-
-    def test_scan_container_redacts_cyclic_dict_without_leaking_original_back_edge(self):
-        payload = {}
-        payload["secret"] = "AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE"
-        payload["self"] = payload
-
-        count, redacted, findings = py_scan_container(
-            payload, {"redact": True, "redaction_text": "[REDACTED]"}
-        )
-
-        assert count == 1
-        assert len(findings) == 1
-        assert redacted["secret"] == "AWS_ACCESS_KEY_ID=[REDACTED]"
-        assert redacted["self"] is redacted
-        assert redacted["self"]["secret"] == "AWS_ACCESS_KEY_ID=[REDACTED]"
-
-    def test_scan_container_redacts_self_referential_object_without_leaking_back_edge(self):
-        class SelfReferentialBox:
-            def __init__(self, secret):
-                self.secret = secret
-                self.self_ref = self
-
-        payload = SelfReferentialBox("AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE")
-
-        count, redacted, findings = py_scan_container(
-            payload, {"redact": True, "redaction_text": "[REDACTED]"}
-        )
-
-        assert count == 1
-        assert len(findings) == 1
-        assert redacted is not payload
-        assert redacted.secret == "AWS_ACCESS_KEY_ID=[REDACTED]"
-        assert redacted.self_ref is redacted
-        assert redacted.self_ref.secret == "AWS_ACCESS_KEY_ID=[REDACTED]"
-
-    def test_scan_container_redacts_tuple_cycle_without_leaking_original_back_edge(self):
-        back_edge = []
-        payload = ("AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE", back_edge)
-        back_edge.append(payload)
-
-        count, redacted, findings = py_scan_container(
-            payload, {"redact": True, "redaction_text": "[REDACTED]"}
-        )
-
-        assert count == 1
-        assert len(findings) == 1
-        assert isinstance(redacted, tuple)
-        assert redacted[0] == "AWS_ACCESS_KEY_ID=[REDACTED]"
-        assert redacted[1][0] is redacted
-        assert redacted[1][0][0] == "AWS_ACCESS_KEY_ID=[REDACTED]"
-
-    def test_scan_container_handles_recursive_model_dump_wrapper_without_crashing(self):
-        script = """
-from cpex_secrets_detection.secrets_detection_rust import py_scan_container
-
-class WrapperPayload:
-    def __init__(self, value):
-        self.value = value
-
-    def model_dump(self):
-        return WrapperPayload(self.value)
-
-count, redacted, findings = py_scan_container(
-    WrapperPayload("AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE"),
-    {"redact": True, "redaction_text": "[REDACTED]"},
-)
-assert count == 1
-assert len(findings) == 1
-assert redacted.value == "AWS_ACCESS_KEY_ID=[REDACTED]"
-"""
-        result = subprocess.run(
-            [sys.executable, "-c", script],
-            capture_output=True,
-            text=True,
-            check=False,
-            timeout=5,
-        )
-
-        assert result.returncode == 0, result.stderr or result.stdout
-
-    def test_scan_container_detects_secret_exposed_only_by_serialized_wrapper_object(self):
-        class View:
-            def __init__(self, secret):
-                self.secret = secret
-
-        class WrappedSerializerModel(BaseModel):
-            safe: str
-
-            @model_serializer(mode="plain")
-            def serialize_model(self):
-                return View("AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE")
-
-        payload = WrappedSerializerModel(safe="ok")
-
-        count, redacted, findings = py_scan_container(
-            payload, {"redact": True, "redaction_text": "[REDACTED]"}
-        )
-
-        assert count == 1
-        assert len(findings) == 1
-        assert isinstance(redacted, View)
-        assert redacted.secret == "AWS_ACCESS_KEY_ID=[REDACTED]"
-
-    def test_scan_container_redacts_self_referential_model_copy_object_without_leak(self):
-        class SelfReferentialModel(BaseModel):
-            secret: str
-            self_ref: "SelfReferentialModel | None" = None
-
-        payload = SelfReferentialModel(secret="AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE")
-        object.__setattr__(payload, "self_ref", payload)
-
-        count, redacted, findings = py_scan_container(
-            payload, {"redact": True, "redaction_text": "[REDACTED]"}
-        )
-
-        assert count == 1
-        assert len(findings) == 1
-        assert redacted is not payload
-        assert isinstance(redacted, SelfReferentialModel)
-        assert redacted.secret == "AWS_ACCESS_KEY_ID=[REDACTED]"
-        assert redacted.self_ref is redacted
-        assert redacted.self_ref.secret == "AWS_ACCESS_KEY_ID=[REDACTED]"
-        assert payload.secret == "AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE"
-        assert payload.self_ref is payload
-
-    def test_scan_container_does_not_call_model_copy_on_clean_object(self):
-        class CountingModel(BaseModel):
-            value: str
-            copies: int = 0
-
-            def model_copy(self, *, update=None, deep=False):
-                object.__setattr__(self, "copies", self.copies + 1)
-                return super().model_copy(update=update, deep=deep)
-
-        payload = CountingModel(value="clean")
-
-        count, redacted, findings = py_scan_container(
-            payload, {"redact": True, "redaction_text": "[REDACTED]"}
-        )
-
-        assert count == 0
-        assert findings == []
-        assert redacted is payload
-        assert payload.copies == 0
-
-    def test_scan_container_detects_secret_exposed_only_by_same_type_serialized_wrapper(self):
-        class SameTypeWrapper(BaseModel):
-            safe: str
-
-            @model_serializer(mode="plain")
-            def serialize_model(self):
-                return SameTypeWrapper.model_construct(
-                    safe="AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE"
-                )
-
-        payload = SameTypeWrapper(safe="clean")
-
-        count, redacted, findings = py_scan_container(
-            payload, {"redact": True, "redaction_text": "[REDACTED]"}
-        )
-
-        assert count == 1
-        assert len(findings) == 1
-        assert isinstance(redacted, SameTypeWrapper)
-        assert redacted.safe == "AWS_ACCESS_KEY_ID=[REDACTED]"
-        assert payload.safe == "clean"
-
-    def test_scan_container_redacts_same_type_serialized_wrapper_without_model_copy(self):
-        class Wrapper:
-            def __init__(self, value):
-                self.value = value
-
-            def model_dump(self):
-                return Wrapper("AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE")
-
-        payload = Wrapper("clean")
-
-        count, redacted, findings = py_scan_container(
-            payload, {"redact": True, "redaction_text": "[REDACTED]"}
-        )
-
-        assert count == 1
-        assert len(findings) == 1
-        assert isinstance(redacted, Wrapper)
-        assert redacted.value == "AWS_ACCESS_KEY_ID=[REDACTED]"
-        assert payload.value == "clean"
-
-    def test_scan_container_handles_recursive_same_type_wrapper_without_rebuild_state(self):
-        script = """
-from cpex_secrets_detection.secrets_detection_rust import py_scan_container
-
-class Wrapper:
-    __slots__ = ()
-
-    def model_dump(self):
-        return Wrapper()
-
-count, redacted, findings = py_scan_container(
-    Wrapper(),
-    {"redact": True, "redaction_text": "[REDACTED]"},
-)
-assert count == 0
-assert findings == []
-assert isinstance(redacted, Wrapper)
-"""
-        result = subprocess.run(
-            [sys.executable, "-c", script],
-            capture_output=True,
-            text=True,
-            check=False,
-            timeout=5,
-        )
-
-        assert result.returncode == 0, result.stderr or result.stdout
-
-    def test_scan_container_handles_tuple_rewrite_with_cyclic_dict_subgraph(self):
-        script = """
-from cpex_secrets_detection.secrets_detection_rust import py_scan_container
-
-d = {}
-payload = ("AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE", d)
-d["self"] = d
-
-count, redacted, findings = py_scan_container(
-    payload,
-    {"redact": True, "redaction_text": "[REDACTED]"},
-)
-assert count == 1
-assert redacted[0] == "AWS_ACCESS_KEY_ID=[REDACTED]"
-assert redacted[1]["self"] is redacted[1]
-"""
-        result = subprocess.run(
-            [sys.executable, "-c", script],
-            capture_output=True,
-            text=True,
-            check=False,
-            timeout=5,
-        )
-
-        assert result.returncode == 0, result.stderr or result.stdout
-
-    def test_scan_container_redacts_slots_declared_as_custom_iterable(self):
-        class SlotNames:
-            def __iter__(self):
-                return iter(("secret",))
-
-        class IterableSlotSecretBox:
-            __slots__ = SlotNames()
-
-            def __init__(self, secret):
-                self.secret = secret
-
-        payload = IterableSlotSecretBox("AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE")
-
-        count, redacted, findings = py_scan_container(
-            payload, {"redact": True, "redaction_text": "[REDACTED]"}
-        )
-
-        assert count == 1
-        assert len(findings) == 1
-        assert isinstance(redacted, IterableSlotSecretBox)
-        assert redacted.secret == "AWS_ACCESS_KEY_ID=[REDACTED]"
-
-
-class TestPluginHookResults:
-    @pytest.fixture
-    def plugin(self):
-        return SecretsDetectionPlugin(_make_config())
-
-    async def test_tool_post_invoke_redacts_mcp_content_payload(self, plugin):
-        payload = ToolPostInvokePayload(
-            name="writer",
-            result={
-                "content": [
-                    {
-                        "type": "text",
-                        "text": "AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE",
-                    }
-                ],
-                "isError": False,
-            },
-        )
-
-        result = await plugin.tool_post_invoke(payload, _make_context())
-
-        assert result.continue_processing is True
-        assert result.modified_payload is not None
-        assert (
-            result.modified_payload.result["content"][0]["text"]
-            == "AWS_ACCESS_KEY_ID=[REDACTED]"
-        )
-        assert result.modified_payload.result["isError"] is False
-        assert result.metadata == {"secrets_redacted": True, "count": 1}
-
-    async def test_tool_post_invoke_preserves_tuple_shape_when_redacted(self, plugin):
-        payload = ToolPostInvokePayload(
-            name="writer",
-            result=("safe", "AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE"),
-        )
-
-        result = await plugin.tool_post_invoke(payload, _make_context())
-
-        assert result.continue_processing is True
-        assert result.modified_payload is not None
-        assert isinstance(result.modified_payload.result, tuple)
-        assert result.modified_payload.result == ("safe", "AWS_ACCESS_KEY_ID=[REDACTED]")
-
-    async def test_prompt_pre_fetch_redacts_nested_custom_object(self, plugin):
-        class SecretBox:
-            def __init__(self, value):
-                self.value = value
-
-        payload = PromptPrehookPayload(
-            prompt_id="prompt-1",
-            args={"payload": SecretBox("AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE")},
-        )
-
-        result = await plugin.prompt_pre_fetch(payload, _make_context())
-
-        assert result.continue_processing is True
-        assert result.modified_payload is not None
-        assert result.modified_payload.args["payload"] is not payload.args["payload"]
-        assert result.modified_payload.args["payload"].value == "AWS_ACCESS_KEY_ID=[REDACTED]"
-        assert payload.args["payload"].value == "AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE"
-
-    async def test_tool_post_invoke_redacts_non_replayable_custom_object(self, plugin):
-        class NonReplayableBox:
-            def __init__(self, secret):
-                self.secret = secret
-                self.derived = "derived"
-
-        payload = ToolPostInvokePayload(
-            name="writer",
-            result=NonReplayableBox("AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE"),
-        )
-
-        result = await plugin.tool_post_invoke(payload, _make_context())
-
-        assert result.continue_processing is True
-        assert result.modified_payload is not None
-        assert result.modified_payload.result is not payload.result
-        assert result.modified_payload.result.secret == "AWS_ACCESS_KEY_ID=[REDACTED]"
-        assert result.modified_payload.result.derived == "derived"
-        assert payload.result.secret == "AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE"
-
-    async def test_tool_post_invoke_redacts_hybrid_dict_and_slots_object(self, plugin):
-        class HybridSecretBox:
-            __slots__ = {"slot_secret": "slot", "__dict__": "dict"}
-
-            def __init__(self, slot_secret, label):
-                self.slot_secret = slot_secret
-                self.label = label
-
-        payload = ToolPostInvokePayload(
-            name="writer",
-            result=HybridSecretBox(
-                "AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE",
-                "safe",
-            ),
-        )
-
-        result = await plugin.tool_post_invoke(payload, _make_context())
-
-        assert result.continue_processing is True
-        assert result.modified_payload is not None
-        assert result.modified_payload.result is not payload.result
-        assert result.modified_payload.result.slot_secret == "AWS_ACCESS_KEY_ID=[REDACTED]"
-        assert result.modified_payload.result.label == "safe"
-        assert payload.result.slot_secret == "AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE"
-
-    async def test_tool_post_invoke_redacts_guarded_object(self, plugin):
-        class GuardedSecretBox:
-            __slots__ = ("secret", "label", "_locked")
-
-            def __init__(self, secret, label):
-                object.__setattr__(self, "secret", secret)
-                object.__setattr__(self, "label", label)
-                object.__setattr__(self, "_locked", True)
-
-            def __setattr__(self, name, value):
-                raise AssertionError(f"unexpected setattr for {name}")
-                object.__setattr__(self, name, value)
-
-        payload = ToolPostInvokePayload(
-            name="writer",
-            result=GuardedSecretBox(
-                "AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE",
-                "safe",
-            ),
-        )
-
-        result = await plugin.tool_post_invoke(payload, _make_context())
-
-        assert result.continue_processing is True
-        assert result.modified_payload is not None
-        assert result.modified_payload.result is not payload.result
-        assert result.modified_payload.result.secret == "AWS_ACCESS_KEY_ID=[REDACTED]"
-        assert result.modified_payload.result.label == "safe"
-
-    async def test_tool_post_invoke_redacts_mapping_slot_object(self, plugin):
-        class MappingSlotSecretBox:
-            __slots__ = {"secret": "slot doc"}
-
-            def __init__(self, secret):
-                self.secret = secret
-
-        payload = ToolPostInvokePayload(
-            name="writer",
-            result=MappingSlotSecretBox("AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE"),
-        )
-
-        result = await plugin.tool_post_invoke(payload, _make_context())
-
-        assert result.continue_processing is True
-        assert result.modified_payload is not None
-        assert result.modified_payload.result is not payload.result
-        assert result.modified_payload.result.secret == "AWS_ACCESS_KEY_ID=[REDACTED]"
-
-    async def test_tool_post_invoke_blocks_secret_exposed_only_by_model_dump(self):
-        class SplitSecretModel(BaseModel):
-            prefix: str
-            suffix: str
-
-            @model_serializer(mode="plain")
-            def serialize_model(self):
-                return f"{self.prefix}{self.suffix}"
-
-        plugin = SecretsDetectionPlugin(_make_config(block_on_detection=True, redact=False))
-        payload = ToolPostInvokePayload(
-            name="writer",
-            result=SplitSecretModel(prefix="AKIA", suffix="FAKE12345EXAMPLE"),
-        )
-
-        result = await plugin.tool_post_invoke(payload, _make_context())
-
-        assert result.continue_processing is False
-        assert result.violation is not None
-        assert result.violation.code == "SECRETS_DETECTED"
-
-    async def test_tool_post_invoke_redacts_secret_exposed_only_by_model_dump(self, plugin):
-        class SplitSecretModel(BaseModel):
-            prefix: str
-            suffix: str
-
-            @model_serializer(mode="plain")
-            def serialize_model(self):
-                return f"{self.prefix}{self.suffix}"
-
-        payload = ToolPostInvokePayload(
-            name="writer",
-            result=SplitSecretModel(prefix="AKIA", suffix="FAKE12345EXAMPLE"),
-        )
-
-        result = await plugin.tool_post_invoke(payload, _make_context())
-
-        assert result.continue_processing is True
-        assert result.modified_payload is not None
-        assert result.modified_payload.result == "[REDACTED]"
-        assert result.metadata == {"secrets_redacted": True, "count": 1}
-
-    async def test_tool_post_invoke_redacts_secret_exposed_only_by_root_model_dump(
-        self, plugin
-    ):
-        payload = ToolPostInvokePayload(
-            name="writer",
-            result=RootModel[str]("AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE"),
-        )
-
-        result = await plugin.tool_post_invoke(payload, _make_context())
-
-        assert result.continue_processing is True
-        assert result.modified_payload is not None
-        assert isinstance(result.modified_payload.result, RootModel)
-        assert result.modified_payload.result.root == "AWS_ACCESS_KEY_ID=[REDACTED]"
-
-    async def test_tool_post_invoke_leaves_clean_payload_unmodified(self, plugin):
-        payload = ToolPostInvokePayload(
-            name="writer",
-            result={
-                "content": [{"type": "text", "text": "plain text"}],
-                "isError": False,
-            },
-        )
-
-        result = await plugin.tool_post_invoke(payload, _make_context())
-
-        assert result.continue_processing is True
-        assert result.violation is None
-        assert result.modified_payload is None
-        assert result.metadata == {}
-
-    async def test_resource_post_fetch_redacts_text_content(self, plugin):
-        payload = ResourcePostFetchPayload(
-            uri="file:///tmp/secret.txt",
-            content=ResourceContent(text="SLACK_TOKEN=xoxr-fake-000000000-fake000000000-fakefakefakefake"),
-        )
-
-        result = await plugin.resource_post_fetch(payload, _make_context())
-
-        assert result.continue_processing is True
-        assert result.modified_payload is not None
-        assert result.modified_payload.content.text == "SLACK_TOKEN=[REDACTED]"
-        assert result.metadata == {"secrets_redacted": True, "count": 1}
-
-    async def test_resource_post_fetch_leaves_clean_payload_unmodified(self, plugin):
-        payload = ResourcePostFetchPayload(
-            uri="file:///tmp/secret.txt",
-            content=ResourceContent(text="plain text"),
-        )
-
-        result = await plugin.resource_post_fetch(payload, _make_context())
-
-        assert result.continue_processing is True
-        assert result.violation is None
-        assert result.modified_payload is None
-        assert result.metadata == {}
-
-    async def test_resource_post_fetch_blocks_when_threshold_met(self):
-        plugin = SecretsDetectionPlugin(
-            _make_config(block_on_detection=True, redact=False, min_findings_to_block=1)
-        )
-        payload = ResourcePostFetchPayload(
-            uri="file:///tmp/secret.txt",
-            content=ResourceContent(text="AWS_ACCESS_KEY_ID=AKIAFAKE12345EXAMPLE"),
-        )
-
-        result = await plugin.resource_post_fetch(payload, _make_context())
-
-        assert result.continue_processing is False
-        assert result.violation is not None
-        assert result.violation.code == "SECRETS_DETECTED"
-        assert result.modified_payload == payload
